// src/models/QuestionDetail.ts
// Generated by https://quicktype.io

export interface QuestionDetail {
  id:              number;
  userId:          string;
  title:           string;
  content:         string;
  score:           number;
  deleted:         boolean;
  createdAt:       string;
  updatedAt:       string;
  facebookId:      string;
  User:            QuestionDetailUser;
  Answer:          Answer[];
  QuestionComment: QuestionComment[];
  QuestionVoter:   Voter[];
  TagsOnQuestions: TagsOnQuestion[];
  facebookLink:    string;
}

export interface Answer {
  id:            number;
  questionId:    number;
  userId:        string;
  content:       string;
  score:         number;
  deleted:       boolean;
  verify:        boolean;
  createdAt:     string;
  updatedAt:     string;
  facebookId:    string;
  AnswerComment: AnswerComment[];
  AnswerVoter:   Voter[];
}

export interface AnswerComment {
  id:         number;
  answerId:   number;
  userId:     string;
  content:    string;
  createdAt:  string;
  updatedAt:  string;
  facebookId: string;
  User:       AnswerCommentUser;
}

export interface AnswerCommentUser {
  uid:         string;
  displayName: string;
  photoURL:    string;
}

export interface Voter {
  userId: string;
  state:  number;
}

export interface QuestionComment {
  id:         number;
  questionId: number;
  userId:     string;
  content:    string;
  createdAt:  string;
  updatedAt:  string;
  facebookId: string;
  User:       QuestionCommentUser;
}

export interface QuestionCommentUser {
  uid:         string;
  displayName: string;
}

export interface TagsOnQuestion {
  Tag: Tag;
}

export interface Tag {
  name: string;
}

export interface QuestionDetailUser {
  uid:         string;
  reputation:  number;
  description: null;
  joiningDate: string;
  displayName: string;
  photoURL:    string;
  phoneNumber: number;
  email:       string;
}