// src/models/Question.ts
// Generated by https://quicktype.io

export interface Question {
    id: number;
    userId: string;
    title: string;
    content: string;
    score: number;
    deleted: boolean;
    createdAt: string;
    updatedAt: string;
    facebookId: string;
    TagsOnQuestions: TagsOnQuestion[];
    User: User;
    _count: Count;
}

export interface TagsOnQuestion {
    Tag: Tag;
}

export interface Tag {
    name: string;
}

export interface User {
    uid: string;
    displayName: string;
    photoURL: string;
}

export interface Count {
    Answer: number;
}